cap: deprecate sticky (=) and ack-required (~) capability modifiers.
There are no known capabilities which use ack-required (~).

Further, it appears that many clients do not properly support the sticky (=) capability modifier,
and in reality, we're required to enforce it with a NAK anyway.  So we just codify the status quo
behaviour to allow clients and servers to drop support for ack-required capabilities and relax the
behaviour of sticky capabilities.