webirc: Fix bad handling of IPv6 addresses
There could potentially be a case when the IPv6 address of a client
begins with a colon (`:`), such as when testing. This could pose an
issue, as the colon symbol is used to denote that the next argument is
the last argument, and should treat all spaces after it as included in
that argument, which means that no further arguments can be passed. This
creates a problem because there is still another argument, `options`,
which needs to be parsed.