Spec cleanups (#441)
Unify spec titles, remove version numbers. Fix duplicate markdown headers.
We've changed filenames and the folder structure for all but Metadata
and SASL because those specs have pending PRs/updates and I don't
want to stomp over them with these.

We've also unified the initial-page-titles for specs.
Mentioned in example-spec.md, but essentially: "Introduction" is for the
first section of a multi-section spec, "Description" is for the main
body section when a spec only contains one section, and "Motivation" is
used when the spec author wants to split out the introduction/description
from the motivation of the spec I guess.
Makes all our spec appear much more nicely on the website, I assure ya'.