Add more FAIL codes for client-initiated batches
While building out my implementation of this, I noticed that many instances of clients sending garbage to the server did not have a defined error code. My implementation provides a baseline support for the BATCH command while allowing other modules to define recognized batch types. As such, the handler for the BATCH command itself is generic and agnostic to the actual type being used. It still needs to do some basic tracking on the supplied reference tag and type before handing it off to the handler module, so cases where the reference tag and/or type are not valid need error messages.

Rather than making something implementation-defined, it'd probably be best to specify some messages that can be used in this case. An implementation that doesn't split things up like mine does isn't required to use these additional messages, but having them specified lets clients know to expect them.